+++ sql-plugin/src/main/320until330-all/scala/com/nvidia/spark/rapids/shims/RapidsParquetScanMeta.scala
@@ -16,7 +16,7 @@
 
 package com.nvidia.spark.rapids.shims
 
-import com.nvidia.spark.rapids.{DataFromReplacementRule, GpuParquetScanBase, RapidsConf, RapidsMeta, ScanMeta}
+import com.nvidia.spark.rapids.{DataFromReplacementRule, GpuParquetScan, RapidsConf, RapidsMeta, ScanMeta}
 
 import org.apache.spark.sql.connector.read.{Scan, SupportsRuntimeFiltering}
 import org.apache.spark.sql.execution.datasources.v2.parquet.ParquetScan
@@ -29,7 +29,7 @@ class RapidsParquetScanMeta(
   extends ScanMeta[ParquetScan](pScan, conf, parent, rule) {
 
   override def tagSelfForGpu(): Unit = {
-    GpuParquetScanBase.tagSupport(this)
+    GpuParquetScan.tagSupport(this)
     // we are being overly cautious and that Parquet does not support this yet
     if (pScan.isInstanceOf[SupportsRuntimeFiltering]) {
       willNotWorkOnGpu("Parquet does not support Runtime filtering (DPP)" +
